move converter under qobj
