+++ qiskit/qobj/converters/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2019, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+"""
+Helper modules to convert qiskit frontend object to proper qobj model.
+"""
